# api.yaml
openapi: 3.0.3
info:
  title: Art Explorer API
  description: API para la aplicación de exploración de arte, gestión de colecciones de usuarios y catálogo de obras.
  version: 1.0.0
servers:
  - url: https://api.artexplorer.com
    description: Servidor de Producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    webhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
      description: Secret key for validating Supabase webhook requests
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: '550e8400-e29b-41d4-a716-446655440000'
...
    CategoryWithRelations:
      type: object
      description: 'Categoría con su jerarquía de relaciones (padres e hijos)'
      required:
        - id
        - name
        - type
        - icon
        - sortOrder
        - parentIds
        - childIds
        - updatedAt
      properties:
        id:
          type: string
          example: 'pintura'
          description: 'ID único de la categoría'
        name:
          type: string
          description: 'Nombre localizado de la categoría según el parámetro language'
          example: 'Pintura'
        type:
          type: string
          enum: [category, subcategory]
          example: 'category'
          description: "Tipo de categoría: 'category' (principal) o 'subcategory' (período artístico)"
        icon:
          type: string
          description: 'Emoji que representa la categoría'
          example: '🎨'
        imageUrl:
          type: string
          nullable: true
          description: 'URL opcional de imagen representativa'
          example: null
        sortOrder:
          type: integer
          example: 1
          description: 'Orden de visualización (menor = primero)'
        parentIds:
          type: array
          items:
            type: string
          description: 'IDs de las categorías padre (vacío para categorías principales)'
          example: []
        childIds:
          type: array
          items:
            type: string
          description: 'IDs de las subcategorías hijas (vacío para subcategorías)'
          example:
            [
              'antiguedad-romana-griega',
              'renacimiento',
              'barroco',
              'impresionismo-postimpresionismo',
            ]
        updatedAt:
          type: string
          format: date-time
          example: '2025-12-04T10:30:00Z'
          description: 'Última actualización para sincronización incremental'

    # --- Home Screen ---
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Invalid image format'
        message:
          type: string
          example: 'The uploaded file must be a valid JPG, PNG, or HEIC image'
...
    User:
      type: object
      description: 'Datos del perfil del usuario'
      required:
        - id
        - email
        - preferredLanguage
        - isPremium
        - createdAt
      properties:
        id:
          $ref: '#/components/schemas/UUID'
          description: 'UUID del usuario (mismo que en Supabase Auth)'
        email:
          type: string
          format: email
          example: 'user@example.com'
        displayName:
          type: string
          example: 'John Doe'
          nullable: true
          description: 'Nombre visible del usuario'
        preferredLanguage:
          type: string
          enum: [es, en]
          example: 'es'
          description: 'Idioma preferido del usuario (detectado automáticamente al registrarse)'
        isPremium:
          type: boolean
          example: false
          description: 'Indica si el usuario tiene suscripción premium activa'
        allowDataCollection:
          type: boolean
          nullable: true
          example: true
          description: 'Permiso para guardar fotos del usuario para mejorar la IA de reconocimiento (null = no se ha preguntado, true = permitido, false = rechazado)'
        createdAt:
          type: string
          format: date-time
          example: '2025-12-01T20:00:00Z'
          description: 'Fecha de creación de la cuenta'
    UpdateUserRequest:
      type: object
      description: 'Datos para actualizar el perfil del usuario'
      properties:
        displayName:
          type: string
          example: 'Jane Smith'
          nullable: true
        preferredLanguage:
          type: string
          enum: [es, en]
          example: 'en'
          description: 'Cambiar idioma preferido del usuario'
        allowDataCollection:
          type: boolean
          nullable: true
          example: true
          description: 'Permiso para guardar fotos para mejorar la IA de reconocimiento'

    # --- Webhooks ---
...
paths:
...
  /api/sync/users/me:
    get:
      tags:
        - Users
      summary: Obtener perfil del usuario autenticado
      description: |
        Retorna la información del perfil del usuario actual.

        **Uso en la app:**
        - Se llama automáticamente después del login para sincronizar datos con SQLite local
        - Los datos se cachean localmente para acceso offline
        - El backend es la fuente de verdad para el perfil del usuario
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: '550e8400-e29b-41d4-a716-446655440000'
                email: 'user@example.com'
                displayName: 'John Doe'
                preferredLanguage: 'es'
                isPremium: false
                allowDataCollection: null
                createdAt: '2025-12-01T20:00:00Z'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'Unauthorized'
                message: 'No authentication token provided'

    patch:
      tags:
        - Users
      summary: Actualizar perfil del usuario autenticado
      description: |
        Actualiza la información del perfil del usuario actual.

        **Campos actualizables:**
        - `displayName`: Nombre visible del usuario
        - `preferredLanguage`: Idioma preferido (es/en)
        - `allowDataCollection`: Permiso para guardar fotos para mejora de IA (true/false/null)

        **Notas:**
        - Solo se actualizan los campos proporcionados (PATCH parcial)
        - El cambio de idioma afecta el contenido personalizado (quotes, descripciones)
        - La app debe sincronizar los cambios con SQLite local después de actualizar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            examples:
              update_language:
                summary: Cambiar idioma preferido
                value:
                  preferredLanguage: 'en'
              update_profile:
                summary: Actualizar nombre
                value:
                  displayName: 'Jane Smith'
              update_all:
                summary: Actualizar todos los campos
                value:
                  displayName: 'Jane Smith'
                  preferredLanguage: 'en'
                  allowDataCollection: true
      responses:
        '200':
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: '550e8400-e29b-41d4-a716-446655440000'
                email: 'user@example.com'
                displayName: 'Jane Smith'
                preferredLanguage: 'en'
                isPremium: false
                allowDataCollection: true
                createdAt: '2025-12-01T20:00:00Z'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'Validation Error'
                message: "preferredLanguage must be 'es' or 'en'"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'Unauthorized'
                message: 'No authentication token provided'
  /api/sync/users/me/push-token:
    post:
      tags:
        - Users
        - Notifications
      summary: Registrar token de notificaciones push
      description: |
        Registra o actualiza el Expo Push Token del dispositivo actual.

        **Uso en la app:**
        - Se llama automáticamente al autenticarse el usuario
        - Se actualiza si el token cambia
        - Un usuario puede tener múltiples tokens (múltiples dispositivos)

        **Backend debe:**
        - Almacenar el token asociado al usuario
        - Actualizar si ya existe un token para el mismo deviceId
        - Mantener múltiples tokens por usuario (para multi-dispositivo)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - deviceId
                - platform
              properties:
                token:
                  type: string
                  example: 'ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx]'
                  description: 'Expo Push Token obtenido del cliente'
                deviceId:
                  type: string
                  example: 'abc123-device-id'
                  description: 'ID único del dispositivo'
                platform:
                  type: string
                  enum: [ios, android]
                  example: 'ios'
                  description: 'Plataforma del dispositivo'
                timestamp:
                  type: integer
                  example: 1702987200000
                  description: 'Timestamp de registro en milisegundos'
      responses:
        '200':
          description: Token registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tokenId:
                    type: string
                    example: 'token-uuid-123'
                    description: 'ID del token registrado en la base de datos'
        '400':
          description: Token inválido o datos incorrectos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'BadRequest'
                message: 'Invalid push token format'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
        - Notifications
      summary: Eliminar token de notificaciones push
      description: |
        Elimina el token del dispositivo actual.

        **Uso en la app:**
        - Se llama al hacer logout
        - Se llama si el usuario desinstala la app (opcional)

        **Backend debe:**
        - Eliminar el token asociado al deviceId del usuario
        - Dejar de enviar notificaciones a ese token
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: deviceId
          required: true
          schema:
            type: string
          description: 'ID del dispositivo cuyo token se eliminará'
          example: 'abc123-device-id'
      responses:
        '200':
          description: Token eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Token no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sync/users/me/notification-preferences:
    get:
      tags:
        - Users
        - Notifications
      summary: Obtener preferencias de notificaciones
      description: |
        Retorna las preferencias de notificaciones del usuario.

        **Preferencias disponibles:**
        - `dailyArtEnabled`: Si el usuario quiere recibir notificación diaria de arte
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Preferencias de notificaciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  dailyArtEnabled:
                    type: boolean
                    example: true
                    description: 'Si el usuario quiere recibir notificación diaria a las 20:00'
              example:
                dailyArtEnabled: true
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Users
        - Notifications
      summary: Actualizar preferencias de notificaciones
      description: |
        Actualiza las preferencias de notificaciones del usuario.

        **Uso en la app:**
        - Se llama cuando el usuario cambia el toggle en Settings
        - El backend debe actualizar las preferencias en la base de datos
        - El cron job de notificaciones debe respetar estas preferencias
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dailyArtEnabled:
                  type: boolean
                  example: true
                  description: 'Si el usuario quiere recibir notificación diaria'
            example:
              dailyArtEnabled: true
      responses:
        '200':
          description: Preferencias actualizadas exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
              example:
                success: true
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sync/categories:
    get:
      tags:
        - Explore
      summary: Obtener todas las categorías con jerarquía y traducciones
      description: |
        Retorna el catálogo completo de categorías de arte con su estructura jerárquica.

        **Estructura de jerarquía:**
        - 3 categorías principales: Pintura, Escultura, Arquitectura y Monumentos
        - 14 subcategorías (períodos artísticos): Antigüedad, Renacimiento, Barroco, etc.
        - Cada subcategoría pertenece a las 3 categorías principales (42 relaciones totales)

        **Uso en la app:**
        - Sincronización inicial al instalar la app o login
        - Actualización periódica en background sync
        - Filtrado en pantalla Explore por categoría/subcategoría
        - Navegación jerárquica: Category → Subcategory → Artworks

        **Sync incremental:**
        - Usar campo `updatedAt` para detectar cambios
        - Cachear localmente en SQLite para acceso offline
        - Las traducciones se sirven según el parámetro `language`
      parameters:
        - in: query
          name: language
          required: false
          schema:
            type: string
            enum: [es, en]
            default: 'es'
          description: |
            Código ISO 639-1 del idioma para las traducciones de nombres de categorías.
            Si no existe traducción para el idioma solicitado, se devuelve español por defecto.
      responses:
        '200':
          description: Lista completa de categorías con relaciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryWithRelations'
                    description: 'Array con todas las categorías (principales y subcategorías)'
                  updatedAt:
                    type: string
                    format: date-time
                    description: 'Timestamp de la última actualización del catálogo (para sync incremental)'
                    example: '2025-12-04T10:30:00Z'
              examples:
                complete_hierarchy:
                  summary: Ejemplo completo con jerarquía de categorías
                  value:
                    categories:
                      - id: 'pintura'
                        name: 'Pintura'
                        type: 'category'
                        icon: '🎨'
                        imageUrl: null
                        sortOrder: 1
                        parentIds: []
                        childIds:
                          - 'antiguedad-romana-griega'
                          - 'arte-medieval-romanico'
                          - 'arte-medieval-gotico'
                          - 'renacimiento'
                          - 'manierismo'
                          - 'barroco'
                          - 'rococo'
                          - 'neoclasicismo'
                          - 'romanticismo'
                          - 'realismo'
                          - 'impresionismo-postimpresionismo'
                          - 'modernismo'
                          - 'vanguardias'
                          - 'arte-moderno-contemporaneo'
                        updatedAt: '2025-12-04T10:00:00Z'
                      - id: 'escultura'
                        name: 'Escultura'
                        type: 'category'
                        icon: '🗿'
                        imageUrl: null
                        sortOrder: 2
                        parentIds: []
                        childIds:
                          - 'antiguedad-romana-griega'
                          - 'arte-medieval-romanico'
                          - 'arte-medieval-gotico'
                          - 'renacimiento'
                          - 'manierismo'
                          - 'barroco'
                          - 'rococo'
                          - 'neoclasicismo'
                          - 'romanticismo'
                          - 'realismo'
                          - 'impresionismo-postimpresionismo'
                          - 'modernismo'
                          - 'vanguardias'
                          - 'arte-moderno-contemporaneo'
                        updatedAt: '2025-12-04T10:00:00Z'
                      - id: 'arquitectura-monumentos'
                        name: 'Arquitectura y Monumentos'
                        type: 'category'
                        icon: '🏛️'
                        imageUrl: null
                        sortOrder: 3
                        parentIds: []
                        childIds:
                          - 'antiguedad-romana-griega'
                          - 'arte-medieval-romanico'
                          - 'arte-medieval-gotico'
                          - 'renacimiento'
                          - 'manierismo'
                          - 'barroco'
                          - 'rococo'
                          - 'neoclasicismo'
                          - 'romanticismo'
                          - 'realismo'
                          - 'impresionismo-postimpresionismo'
                          - 'modernismo'
                          - 'vanguardias'
                          - 'arte-moderno-contemporaneo'
                        updatedAt: '2025-12-04T10:00:00Z'
                      - id: 'renacimiento'
                        name: 'Renacimiento'
                        type: 'subcategory'
                        icon: '🎭'
                        imageUrl: null
                        sortOrder: 4
                        parentIds:
                          - 'pintura'
                          - 'escultura'
                          - 'arquitectura-monumentos'
                        childIds: []
                        updatedAt: '2025-12-04T10:00:00Z'
                      - id: 'barroco'
                        name: 'Barroco'
                        type: 'subcategory'
                        icon: '👑'
                        imageUrl: null
                        sortOrder: 6
                        parentIds:
                          - 'pintura'
                          - 'escultura'
                          - 'arquitectura-monumentos'
                        childIds: []
                        updatedAt: '2025-12-04T10:00:00Z'
                      - id: 'impresionismo-postimpresionismo'
                        name: 'Impresionismo y Post-Impresionismo'
                        type: 'subcategory'
                        icon: '🌈'
                        imageUrl: null
                        sortOrder: 11
                        parentIds:
                          - 'pintura'
                          - 'escultura'
                          - 'arquitectura-monumentos'
                        childIds: []
                        updatedAt: '2025-12-04T10:00:00Z'
                    updatedAt: '2025-12-04T10:30:00Z'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: 'internal_server_error'
                message: 'Failed to retrieve categories from database'
...